icky XPath/namespace parsing is now Blather-native
We used to catch all iq stanzas and do our own checks to see if they
matched the element name and namespace we were looking for.  Something
like this:

```
	query_node = i.children.find { |v| v.element_name == "query" }
	if query_node.namespace.href ==
		'http://jabber.org/protocol/disco#items'
		...
```

But singpolyma pointed out that this isn't necessary.  So now we use
Blather-native message filtering, and things are much prettier.  And,
more importantly, they are also much more correct.

This diff is fairly small because the largest blocks remain the same
(with the same indentation).  Smaller blocks were moved, de-indented.