fix schema: catapult_num now _jid to reflect usage
The "catapult_num-N" keys have always stored the JID associated with N
so it makes more sense to call this "catapult_jid-N" - this is really
a "JID key", not a "number key" (hence the accompanying variable name
change).  Furthermore, we will be adding another number-indexed table
to the schema soon ("catapult_fwd-N"), which will eventually be used
for forwarding calls, but is initially mostly a placeholder in the DB,
used only for sending verification codes via phone from jmp-register.
It makes much less sense to have "catapult_num-N" in this situation.