don't hang on Exception; instead, log msg and quit
This fix introduces several "C: Use 1 (not 0) spaces for indentation."
RuboCop offenses due to the introduction of begin/end blocks that seem
to be required inside each Ruby proc that needs a rescue section.
There wasn't an obvious RuboCop setting to disable, and it's probably
worth fixing in a better way eventually, so leave them in for now.

We did have to disable the RuboCop check for "rescue Exception", which
is used liberally in this commit.  In our situation, doing so is ok,
since we really do want to catch every single type of exception.  It's
fine to do so since we're exiting every time regardless of the issue.