response output was lost in translation in ff64421
In the pre-refactor code, response.to_s looks something like this:
"#<Net::HTTPOK:0x00000002b8dad8>".  After the refactor in ff64421 the
chosen replacement value (http.response) outputs the entire body of
the response instead, which is more than we're looking to log.  So
remove it entirely ("#<EventMachine::HttpClient:0x00000001d6b810>",
the value of http.to_s, is even less useful than response.to_s was)
and just print the response code on its own in that line now.

This happens to fix an 80-column violation introduced in ff64421 as
well, which we should probably have checked by RuboCop at some point.