passthru to_catapult_possible_spam() RV not ignore
It looks like we've been eating this return value for a while.  In
particular, the line appears to have been added in ebfd3a0 and then
went through a few changes after that.  While it hasn't caused any
real problems, it did prevent us from passing back a proper error in
the "SPAM block" portion of 22a898a.  So we want this anyway, and it
will be used in a subsequent commit as well.