add allowlist support for duplicate message blocks
This builds on 22a898a, allowing specified numbers to bypass the
duplicate message blocking added there.  Thanks to 89598a8 we were
also able to add a proper policy-violation error on the first block -
previously the first blocked message was ignored, and the user could
only tell it wasn't delivered through a lack of delivery receipt (but
only if they had that enabled).  Now that edge case is resolved.