short-circuit fix if sgx-bwmsgsv2 users on same DB
Modified copy of "sgx-catapult users on same DB so need passthru fix":
see https://gitlab.com/soprani.ca/sgx-bwmsgsv2/commit/25bc6ad for
details.  Here is a copy of the commit message (edited to correspond
to what is being committed here):

We run sgx-bwmsgsv2 and sgx-catapult off the same database, which
means that mere existence in the database does not mean that we can
pass on the message directly, as they could be on a different system
(i.e. sgx-catapult instead of sgx-bwmsgsv2) and Cheogram will become
unhappy if we try to pass on messages in this way (plus it's wrong).

As a result, we need to make the check more precise, and specifically
determine that the destination user is a user of sgx-catapult, which
we do by checking for a distinctive characteristic of the credentials.
This should ideally be done in a better way, but since we just have
the two systems for now, it is easy to do this substring check, as all
credentials follow this general pattern.

We know this commit works because we've been running it in production
for a couple months now (* sigh *) and have already made the similar
commit to sgx-bwmsgsv2 per above.